Refactor mattermost
* Split up in different files
* Decrease complexity
* Fix linting issues
